chore(infra): move package v1
