+++ coupon/coupon-infra/src/main/java/dev/be/coupon/infra/redis/v1/CouponV1CacheRepository.java
@@ -1,4 +1,4 @@
-package dev.be.coupon.infra.redis;
+package dev.be.coupon.infra.redis.v1;
 
 import dev.be.coupon.domain.coupon.Coupon;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -8,13 +8,13 @@ import java.time.Duration;
 import java.util.Optional;
 import java.util.UUID;
 
-@Repository
-public class CouponCacheRepository {
+@Repository("CouponV1CacheRepository")
+public class CouponV1CacheRepository {
 
     private final RedisTemplate<String, Coupon> redisTemplate;
-    private static final String COUPON_KEY_PREFIX = "coupon:";
+    private static final String COUPON_KEY_PREFIX = "coupon:v1:meta";
 
-    public CouponCacheRepository(RedisTemplate<String, Coupon> redisTemplate) {
+    public CouponV1CacheRepository(final RedisTemplate<String, Coupon> redisTemplate) {
         this.redisTemplate = redisTemplate;
     }
 
@@ -24,7 +24,7 @@ public class CouponCacheRepository {
         return Optional.ofNullable(cached);
     }
 
-    public void save(Coupon coupon) {
+    public void save(final Coupon coupon) {
         String key = COUPON_KEY_PREFIX + coupon.getId();
         // 해당 쿠폰 캐시의 TTL(Time-To-Live) 값
         redisTemplate.opsForValue().set(key, coupon, Duration.ofMinutes(10));
